@@ -35,30 +35,30 @@
     "url": "https://github.com/motorro/firebase-ai-chat/issues"
   },
   "peerDependencies": {
-    "@motorro/firebase-ai-chat-core": "^10.0.3",
-    "firebase-admin": "^10.0.0 || ^11.0.0 || ^12.1.1 || ^13.0.1",
-    "firebase-functions": "^4.9.0 || ^5.0.1 || 6.1.1",
-    "@google-cloud/vertexai": "^1.9.0"
+    "@motorro/firebase-ai-chat-core": "^10.0.4",
+    "firebase-admin": "^10.0.0 || ^11.0.0 || ^12.1.1 || ^13.1.0",
+    "firebase-functions": "^4.9.0 || ^5.0.1 || 6.3.2",
+    "@google-cloud/vertexai": "^1.9.3"
   },
   "devDependencies": {
-    "@johanblumenberg/ts-mockito": "^1.0.41",
-    "@types/chai": "^4.3.5",
-    "@types/chai-as-promised": "^7.1.5",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.16.3",
-    "@types/sinon": "^17.0.3",
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
-    "@typescript-eslint/parser": "^5.12.0",
-    "chai": "^4.3.7",
-    "chai-as-promised": "^7.1.1",
-    "eslint": "^8.9.0",
+    "@johanblumenberg/ts-mockito": "^1.0.45",
+    "@types/chai": "^5.0.1",
+    "@types/chai-as-promised": "^8.0.1",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.13.5",
+    "@types/sinon": "^17.0.4",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
+    "@typescript-eslint/parser": "^8.25.0",
+    "chai": "^5.2.0",
+    "chai-as-promised": "^8.0.1",
+    "eslint": "^9.21.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-import": "^2.25.4",
-    "firebase-functions-test": "^3.2.0",
-    "firebase-tools": "^13.7.2",
-    "mocha": "^10.2.0",
-    "sinon": "^17.0.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.0"
+    "eslint-plugin-import": "^2.31.0",
+    "firebase-functions-test": "^3.4.0",
+    "firebase-tools": "^13.31.2",
+    "mocha": "^11.1.0",
+    "sinon": "^19.0.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   }
 }
