@@ -1,7 +1,5 @@
-// eslint-disable-next-line
-const chai = require("chai");
-// eslint-disable-next-line
-const chaiAsPromised = require("chai-as-promised");
+import * as chai from 'chai';
+import chaiAsPromised from 'chai-as-promised';
 
-chai.should();
 chai.use(chaiAsPromised);
+chai.should();
